renaming: json_schema_merger -> schema_merger, draft4_generator -> schema_generator
