+++ bin/schema_merger
@@ -10,7 +10,7 @@ import sys
 import os
 sys.path.append(os.path.dirname(os.path.dirname(__file__)))
 
-from skinfer.json_schema_merger import merge_schema
+from skinfer.schema_merger import merge_schema
 import json
