+++ skinfer/schema_generator.py
