+++ tests/test_schema_merger.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 
 from __future__ import absolute_import, print_function
-from skinfer.json_schema_merger import merge_schema
+from skinfer.schema_merger import merge_schema
 import unittest
 from tests import fixtures
